Use CString for port
This shifts a potential panic outside of the library, and avoids the
unnecessary overhead of UTF-8 validation when a port is dynamically
received (eg. through walking the file-system / udev hotplug events).

It should also fix connecting to a device with a path that isn't valid
UTF-8, but I don't think that'd ever happen in practice.